# ENTREGABLES_SIGAI_SES/01_DOCUMENTACION_TECNICA/API_SPEC/openapi.yaml
openapi: 3.0.3
info:
  title: SIGAI-SES API
  version: "1.0.0"
  description: >
    API RESTful del Sistema Integral de Gestión de Activos e Inventario
    para Seguridad Electrónica Securitas Colombia.
  contact:
    name: Equipo de Desarrollo SIGAI-SES
    email: soporte@ses-securitas.com.co
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.ses.example.com/v1
    description: Producción
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido de /auth/login
  schemas:
...
    Token:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
        prioridad:
          type: string
          enum: [critica, alta, media, baja]
        estado:
          type: string
          enum: [activa, reconocida, resuelta, ignorada]
        titulo:
          type: string
        descripcion:
          type: string
        item_nombre:
          type: string
        valor_actual:
          type: number
        valor_umbral:
          type: number
        created_at:
          type: string
          format: date-time

    # ==================== AUDITORIA ====================
...
    Error:
      type: object
      properties:
        detail:
          type: string
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Iniciar sesion
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required: [username, password]
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "401":
          description: Credenciales invalidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      summary: Renovar access token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token renovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
...
  /alerts:
    get:
      summary: Listar alertas
      tags: [Alerts]
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [activa, reconocida, resuelta, ignorada]
        - name: prioridad
          in: query
          schema:
            type: string
            enum: [critica, alta, media, baja]
        - name: tipo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      summary: Crear alerta manual
      tags: [Alerts]
      responses:
        "201":
          description: Alerta creada
...
  /alerts/{id}/estado:
    patch:
      summary: Actualizar estado de alerta
      tags: [Alerts]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                estado:
                  type: string
                  enum: [activa, reconocida, resuelta, ignorada]
      responses:
        "200":
          description: Estado actualizado

  # ==================== REGIONALES ====================
...